Phase 5 commit 7: extract FolderSidebar.tsx
The sidebar's tree renderer (renderTreeNode) and filter machinery
(filterMatchesNode, renderFilteredNodes) move with it, plus the
DRAG_MIME constant + NoteDragPayload type the sidebar drag/drop
needs.

FolderSidebar accepts ~25 props from Notes (all UI state, setters,
and handlers). The split keeps Notes.tsx's file-ops handlers
unchanged; the sidebar just delegates to them via callback props.

Notes.tsx 963 → 731 lines (was 1108 before Phase 5).

Build clean, tests baseline unchanged.
